Also drop resource link target FK
